+++ Aplicatie/Water Monitoring System - Android App/app/src/main/java/com/example/watermonitoringsystem/activities/common/SensorsMapActivity.java
@@ -1,4 +1,4 @@
-package com.example.watermonitoringsystem.activities.supplier;
+package com.example.watermonitoringsystem.activities.common;
 
 import static com.example.watermonitoringsystem.authentication.LogoutHelper.logoutFromActivity;
 
@@ -26,9 +26,12 @@ import androidx.core.view.GravityCompat;
 import androidx.drawerlayout.widget.DrawerLayout;
 
 import com.example.watermonitoringsystem.R;
-import com.example.watermonitoringsystem.activities.common.AboutAppActivity;
-import com.example.watermonitoringsystem.activities.common.AppSupportActivity;
-import com.example.watermonitoringsystem.activities.common.SensorsModuleInfoActivity;
+import com.example.watermonitoringsystem.activities.customer.CustomerComplaintsActivity;
+import com.example.watermonitoringsystem.activities.customer.CustomerDashboardActivity;
+import com.example.watermonitoringsystem.activities.customer.CustomerPersonalProfileActivity;
+import com.example.watermonitoringsystem.activities.supplier.AddCoordinateToExistingSensor;
+import com.example.watermonitoringsystem.activities.supplier.SupplierElectrovalveActivity;
+import com.example.watermonitoringsystem.activities.supplier.SupplierWaterPumpActivity;
 import com.example.watermonitoringsystem.api.ApiManager;
 import com.example.watermonitoringsystem.authentication.SharedPrefsKeys;
 import com.example.watermonitoringsystem.firebase.Database;
@@ -39,6 +42,7 @@ import com.example.watermonitoringsystem.models.sqldb.AllSensorsRealTimeDataResp
 import com.example.watermonitoringsystem.models.sqldb.RegisteredRawElementsData;
 import com.example.watermonitoringsystem.models.sqldb.RegisteredElementData;
 import com.example.watermonitoringsystem.mqtt.MqttConstants;
+import com.example.watermonitoringsystem.utils.Constants;
 import com.example.watermonitoringsystem.utils.Utils;
 import com.google.android.gms.maps.CameraUpdate;
 import com.google.android.gms.maps.CameraUpdateFactory;
@@ -50,6 +54,7 @@ import com.google.android.gms.maps.model.LatLng;
 import com.google.android.gms.maps.model.LatLngBounds;
 import com.google.android.gms.maps.model.Marker;
 import com.google.android.gms.maps.model.MarkerOptions;
+import com.google.android.material.appbar.AppBarLayout;
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
 import com.google.android.material.navigation.NavigationView;
 import com.google.firebase.database.DataSnapshot;
@@ -73,13 +78,15 @@ import retrofit2.Response;
  *
  * @author Ioan-Alexandru Chirita
  */
-public class SupplierSensorsMapActivity extends AppCompatActivity implements NavigationView.OnNavigationItemSelectedListener, OnMapReadyCallback {
+public class SensorsMapActivity extends AppCompatActivity implements NavigationView.OnNavigationItemSelectedListener, OnMapReadyCallback {
 
-    private static final String TAG = SupplierSensorsMapActivity.class.getSimpleName();
-    private static final String MAP_CAMERA_POS_KEY = "com.example.watermonitoringsystem.activities.supplier.SupplierSensorsMapActivity.map_camera_pos";
+    private static final String TAG = SensorsMapActivity.class.getSimpleName();
+    private static final String MAP_CAMERA_POS_KEY = "com.example.watermonitoringsystem.activities.common.SupplierSensorsMapActivity.map_camera_pos";
     private static final String CAMERA_LAT_KEY = "camera_lat_key";
     private static final String CAMERA_LON_KEY = "camera_lon_key";
     private static final String CAMERA_ZOOM_KEY = "camera_zoom_key";
+    private String userType;
+    private String customerCode;
     private boolean addSensorMode;
     ActivityResultLauncher<Intent> launchSensorModuleInfo;
     private ArrayList<SensorData> sensorDataListForMap;
@@ -96,6 +103,9 @@ public class SupplierSensorsMapActivity extends AppCompatActivity implements Nav
         super.onCreate(savedInstanceState);
 
         setContentView(R.layout.supplier_sensors_activity);
+
+        userType = Utils.getValueFromSharedPreferences(SharedPrefsKeys.KEY_USER_TYPE, this);
+        customerCode = Utils.getValueFromSharedPreferences(SharedPrefsKeys.KEY_CUSTOMER_CODE, this);
 
         launchSensorModuleInfo = registerForActivityResult(
                 new ActivityResultContracts.StartActivityForResult(),
@@ -121,17 +131,11 @@ public class SupplierSensorsMapActivity extends AppCompatActivity implements Nav
         );
 
         // Toolbar
-        Toolbar toolbar = findViewById(R.id.toolbar_supplier);
-        setSupportActionBar(toolbar);
-        Objects.requireNonNull(getSupportActionBar()).setDisplayShowTitleEnabled(false);
-        Objects.requireNonNull(getSupportActionBar()).setHomeButtonEnabled(true);
+        AppBarLayout appBarLayout = findViewById(R.id.appBar);
+        Toolbar mainToolbar;
+        Toolbar supplierToolbar = appBarLayout.findViewById(R.id.toolbar_supplier);
+        Toolbar customerToolbar = appBarLayout.findViewById(R.id.toolbar_customer);
 
-        // Drawer Layout
-        DrawerLayout drawer = findViewById(R.id.drawer_layout_supplier_sensors);
-        ActionBarDrawerToggle toggle = new ActionBarDrawerToggle(
-                this, drawer, toolbar, R.string.navigation_drawer_open, R.string.navigation_drawer_close);
-        drawer.addDrawerListener(toggle);
-        toggle.syncState();
 
         // Navigation View
         NavigationView navigationView = findViewById(R.id.nav_view_supplier);
@@ -139,27 +143,55 @@ public class SupplierSensorsMapActivity extends AppCompatActivity implements Nav
         navigationView.setCheckedItem(R.id.nav_sensors);
         View headerLayout = navigationView.getHeaderView(0);
 
+        // Update Toolbar and NavigationView regarding to the user type = SUPPLIER / CUSTOMER
+        if (userType.equals(Constants.SUPPLIER)) {
+            // Remove toolbar for customer and set toolbar for supplier
+            appBarLayout.removeView(customerToolbar);
+            mainToolbar = supplierToolbar;
+            // Inflate supplier menu
+            navigationView.inflateMenu(R.menu.navigation_drawer_supplier);
+
+            // Get notifications number
+            ImageView redSquare = findViewById(R.id.red_square);
+            TextView notificationNumber = findViewById(R.id.notifications_number);
+            Utils.getNotificationsNumber(redSquare, notificationNumber);
+
+            // Notification Bell
+            ImageView notificationsBellImg = findViewById(R.id.notificationBellImg);
+            notificationsBellImg.setOnClickListener(Utils::openNotificationsPopUp);
+        } else {
+            // Remove toolbar for supplier and set toolbar for customer
+            appBarLayout.removeView(supplierToolbar);
+            mainToolbar = customerToolbar;
+            // Inflate customer menu
+            navigationView.inflateMenu(R.menu.navigation_drawer_customer);
+        }
+
+        // Toolbar
+        setSupportActionBar(mainToolbar);
+        Objects.requireNonNull(getSupportActionBar()).setDisplayShowTitleEnabled(false);
+        Objects.requireNonNull(getSupportActionBar()).setHomeButtonEnabled(true);
+
+        // Drawer Layout
+        DrawerLayout drawer = findViewById(R.id.drawer_layout_supplier_sensors);
+        ActionBarDrawerToggle toggle = new ActionBarDrawerToggle(
+                this, drawer, mainToolbar, R.string.navigation_drawer_open, R.string.navigation_drawer_close);
+        drawer.addDrawerListener(toggle);
+        toggle.syncState();
+
         TextView txtName = headerLayout.findViewById(R.id.user_nav_header);
         TextView txtEmail = headerLayout.findViewById(R.id.email_nav_header);
         CircleImageView imgProfile = headerLayout.findViewById(R.id.profile_picture_nav_header);
 
-        String email = Utils.getValueFromSharedPreferences(SharedPrefsKeys.KEY_EMAIL, SupplierSensorsMapActivity.this);
+        String email = Utils.getValueFromSharedPreferences(SharedPrefsKeys.KEY_EMAIL, SensorsMapActivity.this);
         Utils.getSupplierProfileFromDatabase(email, txtName, txtEmail, imgProfile);
 
-        // Get notifications number
-        ImageView redSquare = findViewById(R.id.red_square);
-        TextView notificationNumber = findViewById(R.id.notifications_number);
-        Utils.getNotificationsNumber(redSquare, notificationNumber);
-
-        // Notification Bell
-        ImageView notificationsBellImg = findViewById(R.id.notificationBellImg);
-        notificationsBellImg.setOnClickListener(Utils::openNotificationsPopUp);
 
         // Create Google Map
         mMarkerArray = new ArrayList<>();
         SupportMapFragment mapFragment = (SupportMapFragment) getSupportFragmentManager().findFragmentById(R.id.map);
         assert mapFragment != null;
-        mapFragment.getMapAsync(SupplierSensorsMapActivity.this);
+        mapFragment.getMapAsync(SensorsMapActivity.this);
 
         // Add sensor Text View
         addSensorTextView = findViewById(R.id.addSensorTextView);
@@ -191,6 +223,11 @@ public class SupplierSensorsMapActivity extends AppCompatActivity implements Nav
             }
             buildGoogleMapsWithMarkers(googleMap);
         });
+
+        if (userType.equals(Constants.CUSTOMER)) {
+            toggleViewFab.setVisibility(View.GONE);
+            addSensorFab.setVisibility(View.GONE);
+        }
     }
 
 
@@ -211,17 +248,26 @@ public class SupplierSensorsMapActivity extends AppCompatActivity implements Nav
     public boolean onNavigationItemSelected(MenuItem item) {
         int id = item.getItemId();
 
-        if (id == R.id.nav_electrovalve) {
+        if (id == R.id.nav_sensors) {
+            startActivity(new Intent(this, SensorsMapActivity.class));
+            finish();
+        } else if (id == R.id.nav_electrovalve) {
             startActivity(new Intent(this, SupplierElectrovalveActivity.class));
             finish();
         } else if (id == R.id.nav_water_pump) {
             startActivity(new Intent(this, SupplierWaterPumpActivity.class));
             finish();
+        } else if (id == R.id.nav_home_customer) {
+            startActivity(new Intent(this, CustomerDashboardActivity.class));
+            finish();
+        } else if (id == R.id.nav_personal_data) {
+            startActivity(new Intent(this, CustomerPersonalProfileActivity.class));
+            finish();
+        } else if (id == R.id.nav_complaints) {
+            startActivity(new Intent(this, CustomerComplaintsActivity.class));
+            finish();
         } else if (id == R.id.nav_app_support) {
             startActivity(new Intent(this, AppSupportActivity.class));
-            finish();
-        } else if (id == R.id.nav_about_app) {
-            startActivity(new Intent(this, AboutAppActivity.class));
             finish();
         } else if (id == R.id.nav_sign_out) {
             logoutFromActivity(this);
@@ -395,6 +441,8 @@ public class SupplierSensorsMapActivity extends AppCompatActivity implements Nav
         mMarkerArray.clear();
         googleMap.clear();
         for (SensorData sensor : sensorDataListForMap) {
+            if(userType.equals(Constants.CUSTOMER) && !sensor.getCustomerCode().equals(customerCode))
+                continue;
             if(!viewFlag && !sensor.hasDataChannels())
                 continue;
             LatLng place = new LatLng(sensor.getLatitude(), sensor.getLongitude());
@@ -415,7 +463,7 @@ public class SupplierSensorsMapActivity extends AppCompatActivity implements Nav
         googleMap.setOnMapClickListener(latLng -> {
             if(!addSensorMode)
                 return;
-            new AlertDialog.Builder(SupplierSensorsMapActivity.this).setTitle("Add new location")
+            new AlertDialog.Builder(SensorsMapActivity.this).setTitle("Add new location")
                 .setMessage("Are you sure you want to add a new location and a customer code for an existing sensor?")
                 .setPositiveButton("Yes", (dialog, which) -> {
                     Bundle b = new Bundle();
@@ -423,7 +471,7 @@ public class SupplierSensorsMapActivity extends AppCompatActivity implements Nav
                     DecimalFormat numberFormat = new DecimalFormat("#.000000");
                     b.putDouble(getString(R.string.latitude_field), Double.parseDouble(numberFormat.format(latLng.latitude)));
                     b.putDouble(getString(R.string.longitude_field), Double.parseDouble(numberFormat.format(latLng.longitude)));
-                    Intent intent = new Intent(SupplierSensorsMapActivity.this, AddCoordinateToExistingSensor.class);
+                    Intent intent = new Intent(SensorsMapActivity.this, AddCoordinateToExistingSensor.class);
                     intent.putExtras(b);
                     startActivity(intent);
                     //finish();
@@ -436,7 +484,7 @@ public class SupplierSensorsMapActivity extends AppCompatActivity implements Nav
      * Open sensor's channel activity
      */
     private void openSensorModuleInfo(String moduleId, String customerCode) {
-        Intent intent = new Intent(SupplierSensorsMapActivity.this, SensorsModuleInfoActivity.class);
+        Intent intent = new Intent(SensorsMapActivity.this, SensorsModuleInfoActivity.class);
         Bundle b = new Bundle();
         b.putString(getString(R.string.sensor_id_field), moduleId);
         b.putString(getString(R.string.customer_code_field), customerCode);